+++ packages/awino-ui/src/components/general/AssetAmount/AssetAmount.test.tsx
@@ -1,9 +1,11 @@
 import React from 'react';
 
+import BigNumber from 'bignumber.js';
+
 import nextUseRouterMock from '@/mocks/nextUseRouterMock';
 import { cleanup, render } from '@/testing/utils';
 
-import PurchaseSection from './PurchaseSection';
+import AssetAmount from './AssetAmount';
 
 beforeAll(() => {
   nextUseRouterMock({
@@ -16,9 +18,11 @@ beforeAll(() => {
 
 afterEach(cleanup);
 
-describe('<PurchaseSection />', () => {
+describe('<AssetAmount />', () => {
   it('has valid snapshot', () => {
-    const { asFragment } = render(<PurchaseSection />, {}, 'podl');
+    const { asFragment } = render(
+      <AssetAmount asset="ftm" value={new BigNumber(1234567.89)} altAsset="usd" altValue={new BigNumber(9876543.21)} />
+    );
     expect(asFragment()).toMatchSnapshot();
   });
 });