+++ packages/awino-ui/src/components/general/SwappingImage/SwappingImage.test.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import nextUseRouterMock from '@/mocks/nextUseRouterMock';
 import { cleanup, render } from '@/testing/utils';
 
-import PurchaseSection from './PurchaseSection';
+import SwappingImage from './SwappingImage';
 
 beforeAll(() => {
   nextUseRouterMock({
@@ -16,9 +16,9 @@ beforeAll(() => {
 
 afterEach(cleanup);
 
-describe('<PurchaseSection />', () => {
+describe('<SwappingImage />', () => {
   it('has valid snapshot', () => {
-    const { asFragment } = render(<PurchaseSection />, {}, 'podl');
+    const { asFragment } = render(<SwappingImage source="ftm" target="awi" />);
     expect(asFragment()).toMatchSnapshot();
   });
 });