+++ packages/awino-ui/src/components/pages/podl/PurchaseSection/Panel.tsx
@@ -13,7 +13,7 @@ const Panel = styled(Box)(({ theme }) => ({
     justifyContent: 'space-between',
     gap: theme.spacing(10),
     padding: theme.spacing(5.5, 6.5, 5),
-    margin: theme.spacing(0, 0, 15, 0),
+    margin: theme.spacing(0, 0, 10, 0),
     borderRadius: +theme.shape.borderRadius * 5,
     backgroundColor: theme.palette.background.transparent,
     '.aside': {
@@ -22,7 +22,7 @@ const Panel = styled(Box)(({ theme }) => ({
     },
   },
   '.content': {
-    padding: theme.spacing(4, 12.5, 10),
+    padding: theme.spacing(4, 5, 10),
     '.table-container': {
       height: 888 /* 66 * 10 + 12 * 10 - 12 */,
       width: '100%',
@@ -34,6 +34,9 @@ const Panel = styled(Box)(({ theme }) => ({
       alignItems: 'center',
       gap: theme.spacing(18),
       padding: theme.spacing(5.5, 12.5, 5),
+    },
+    '.content': {
+      padding: theme.spacing(4, 12.5, 10),
     },
   },
 }));