+++ packages/awino-ui/src/pages/podl.tsx
@@ -8,23 +8,11 @@ import storeWrapper from '@/app/store';
 import Seo from '@/components/layout/Seo/Seo';
 import PurchaseSection from '@/components/pages/podl/PurchaseSection/PurchaseSection';
 
-// const stats = [
-//   { value: 89.7, subvalue: 24.72 },
-//   { value: 89.7, subvalue: 24.72 },
-//   { value: 0.27 },
-//   { value: 273.4, subvalue: 52 },
-//   { value: 47.4, subvalue: 24.72 },
-// ];
-
-// const assets = { market: 12345678, platform: 1234567.89 };
-
 const PodlPage: NextPage = () => {
   return (
     <>
       <Seo />
       <PurchaseSection />
-      {/* <StatsSection items={stats} /> */}
-      {/* <AssetSection total={assets} /> */}
     </>
   );
 };